+++ app/javascript/packs/application.js
@@ -1,6 +1,6 @@
-require ('bootstrap')
-require('@fortawesome/fontawesome-free/js/all')
-require( '../src/application.scss')
+import 'bootstrap'
+import '@fortawesome/fontawesome-free/js/all'
+import '../src/application.scss'
 require("@rails/ujs").start()
 require("turbolinks").start()
 require("@rails/activestorage").start()
